Change: use automatic gradient
This uses automatic gradients, making the `gradient` parameter optional.
Gradients are determined directly from the passed `J_T`, either by
selecting the appropriate implementation of the analytic gradient, or by
constructing a routine via automatic differentiation (via Zygote).

In addition, we can evaluate ∇J_T not just via a chain rules with
respect to overlaps τ_k as we have done so far, but also via a chain
rule with respect to the forward-propagated states. This allows to set
automatic gradients for functionals that cannot be written in terms of
overlaps, and/or of objectives that do not define a `target_state`.

This required a major rewrite of the core algorithm: Before, we would
backward-propagate the target states, and the forward-propagate the
combined initial states and gradients. This is not compatible with
automatically evaluating gradients via a chain rule in the states. In
that case, we have to backward-propagate not the target states, but
states χ=-∂J/∂⟨ϕ| that depend on the forward-propagated states. Thus, we
have to flip the order of propagations: first forward-propagate that
initial states, and then backward-propagate the combined χ-states and
gradients.

This implements a basic "semi-automatic differentiation" for GRAPE.