Implement L-BFGS-B bounds
the `optimize_grape` routine can now be called with an `upper_bound` and
`lower_bound` parameter. For time-dependent bounds, or different bounds
on different pulses, it is also possible to give `pulse_options` with
keys `upper_bounds` and `lower_bounds`.

This is only implemented for the L-BFGS-B backend. I couln't really
figure out how to add bounds to Optim.jl.

https://github.com/JuliaQuantumControl/QuantumControl.jl/issues/15