Refactor LBFGSB code into `ext` folder
This is a "pretend" package extension, for symmetry with the Optim
extension. It's not possible to have a package extension triggered by a
hard dependency, so this uses `include` instead.

However, if we ever wanted to switch the default backend, this makes
that easy, as we'd only have to tweak the import in `src/GRAPE.jl` and
the list of `[extensions]` in the main `Project.toml`.